Fix compilation regression (for #1)
This change conflates the distinction introduced in Play JSON 2.4 between values and lookup results, and should be considered temporary. There's more work ahead to determine if `Node` should represent JsLookupResult instead. (Which appears to @mandubian's original intent with his use of JsUndefined for the Empty and Error nodes.) See discussion on issue #1 for more details.